Why is Pacman yellow?
In an interview, the creator Toru Iwatani said that the inspiration
for Pac-Man was a pizza with a slice missing, and Pac-Man is meant to
be an animated pizza, with a missing slice as a mouth. The yellow
color would then be chosen from the yellow cheese of the pizza. He
certainly is a beloved character, but Japan is weird.